requirements.txt
Flask==3.1.3
Pillow==12.1.1
Werkzeug==3.1.5
